requirements.txt
openai>=1.50
numpy>=1.24
Pillow>=10.0
opencv-python-headless>=4.8
omegaconf>=2.3
tqdm>=4.66
plyfile>=1.0
scipy>=1.11
